Fix Issue with multiple calls to close()
As it is now if close() is called more than once per show (or called
before show) some of the class attributes will not exist during the
following calls to close. Added check to see if self.window has been
setup to detect a previous call to close().  This seems to be happening
in: hsoft/dupeguru#449.